Add login requireq to m decorator to product modification

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 from .models import Product, Category, Philosopher
@@ -44,7 +45,7 @@ def all_products(request):
             products = products.filter(philosopher__name__in=philosophers)
             philosophers = Philosopher.objects.filter(name__in=philosophers)
 
-        
+
         if 'q' in request.GET:
             query = request.GET['q']
             if not query:
@@ -79,8 +80,13 @@ def product_detail(request, product_id):
     return render(request, 'products/product_detail.html', context)
 
 
+@login_required()
 def add_product(request):
     """ Add a product to the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only the site admin can access this page.')
+        return redirect (reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -99,9 +105,14 @@ def add_product(request):
 
     return render(request, template, context)
 
-    
+
+@login_required()
 def edit_product(request, product_id):
     """ Edit a product in the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only the site admin can access this page.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
 
     if request.method == 'POST':
@@ -125,8 +136,13 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required()
 def delete_product(request, product_id):
     """ Delete a product in the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only the site admin can access this page.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(request, pk=product_id)
     product.delete()
     messages.success(request, 'Product successfully deleted!')
